StreamSize uint16_t > uint32_t for large file support
* StreamSize was small for embedded use-cases but most modern systems can eentually exceed this even in embedded use-cases as it is not uncommon to have >64KB ram

+++ include/sub0pub/sub0pub.hpp
@@ -140,7 +140,7 @@ namespace sub0
         class OStream
         {
         public:
-            typedef uint_fast16_t StreamSize;
+            typedef uint_fast32_t StreamSize;
 
             virtual StreamSize write(const char* const buffer, const StreamSize bufferCount) = 0;
 
@@ -155,7 +155,7 @@ namespace sub0
         class IStream
         {
         public:
-            typedef uint_fast16_t StreamSize;
+            typedef uint_fast32_t StreamSize;
 
             virtual StreamSize read(char* const buffer, const StreamSize bufferCount) = 0;
